+++ config/initializers/container.rb
@@ -67,11 +67,12 @@ module Charkeeper
 
     register('commands.characters_context.dnd2024.create') { CharactersContext::Dnd2024::CreateCommand.new }
     register('commands.characters_context.dnd2024.update') { CharactersContext::Dnd2024::UpdateCommand.new }
-    register('commands.characters_context.dnd2024.spell_update') { CharactersContext::Dnd2024::SpellUpdateCommand.new }
-    register('commands.characters_context.dnd2024.spell_add') { CharactersContext::Dnd2024::SpellAddCommand.new }
     register('commands.characters_context.dnd2024.make_short_rest') { CharactersContext::Dnd2024::MakeShortRestCommand.new }
     register('commands.characters_context.dnd2024.make_long_rest') { CharactersContext::Dnd2024::MakeLongRestCommand.new }
     register('commands.characters_context.dnd2024.craft') { CharactersContext::Dnd2024::CraftCommand.new }
+
+    register('commands.characters_context.dnd2024.spells.change') { CharactersContext::Dnd2024::Spells::ChangeCommand.new }
+    register('commands.characters_context.dnd2024.spells.add') { CharactersContext::Dnd2024::Spells::AddCommand.new }
 
     register('commands.characters_context.dnd2024.talents.add') { CharactersContext::Dnd2024::Talents::AddCommand.new }